Add files via upload
I have worked on Predicting House Prices Project using Machine Learning. In this task, we performed a series of steps to develop a linear regression model for predicting house prices based on a dataset. First, we loaded and explored the dataset to ensure there were no missing values. We selected three key features: the size of the house in square feet, the number of bedrooms, and the number of bathrooms. These features were used to train the model. We Have create a bar graph to visualize the relationship between one of the features and the target variable. We have created a line graph between average house price against both the size of the house (in square feet) and the number of bedrooms. We then split the data into training and testing sets to evaluate the model's performance. A linear regression model was trained using the training set, and predictions were made on the test set. Finally, we evaluated the model using the Root Mean Squared Error (RMSE) metric, which provided an indication of the average prediction error in dollars.