Add files via upload
I have completed my project of Classifying Emails as Spam or Not Spam. I have taken the dataset from Kaggle and  code processes an email dataset (Email_data.csv) for classification tasks. It starts by checking for missing values and examining the distribution of the target variable, 'Prediction'. Feature extraction is performed using TF-IDF on the email content, transforming the text into numerical data. The data is then split into training and testing sets, and two models, Naive Bayes and SVM, are trained on the data. Their performance is evaluated using accuracy scores, classification reports, and confusion matrices, with results visualized through count plots and heatmaps. PCA is applied for dimensionality reduction to visualize the data in 2D, and a heatmap is generated to explore feature correlations.